split filter code from calibrate.py to new filter.py
