+++ calibrate/calibrate.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 """
-Library and cli for motor calibration
+Library for motor calibration
 TODO:
-    more digital filter tuning (IIR vs FIR?)
     slowness:
         look at slowness factor (coulomb (and/or static?) fric too high)
         graph slowness factor and acceleration?
@@ -17,7 +16,6 @@ TODO:
                 ex "Vsq": (lambda results: return np.abs(results["V_f"]) * results["V_f"])
     graph individual variable contributions to model (use X matrix)
     think more about inductance (can model with 3rd derivative of angle)
-    way to run with saved parameters
     electrically isolate angle sensors to reduce noise?
 
 Created Jun 2021
@@ -25,30 +23,9 @@ Created Jun 2021
 """
 
 import numpy as np
-from scipy import signal
 from scipy.fft import rfftfreq
-from dataclasses import dataclass
 import time
-
-@dataclass
-class FilterParams:
-    T: int = None
-    N: int = 60
-    order: int = 6
-    fcn: str = "butter"   # 'cheby'
-    ripple: float = 0.5  # only applies to cheby filters
-    def __post_init__(self):
-        self.set_N(self.N if self.N is not None else 60)
-    def set_N(self, N: int | None, t_data: np.ndarray = None):
-        if N is None:
-            N = np.argmax(t_data > self.T) - 1
-        self.N = N
-        if self.fcn == 'cheby':
-            self.lowpass_sos = signal.cheby1(self.order, self.ripple, 1/N, 'lowpass',  output='sos')
-            self.hipass_sos  = signal.cheby1(self.order, self.ripple, 1/N, 'highpass', output='sos')
-        else:
-            self.lowpass_sos = signal.butter(self.order, 1/N, 'lowpass', output='sos')
-            self.hipass_sos  = signal.butter(self.order, 1/N, 'highpass', output='sos')
+from filter import *
 
 
 def examineMotor(testdata, model=None, params=None, filter_params=FilterParams()):
@@ -95,18 +72,19 @@ def examineMotor(testdata, model=None, params=None, filter_params=FilterParams()
         V_f = results["V_f"]
         vel_f = results["vel_f"]
         X = np.hstack((V_f.reshape(-1,1), vel_f.reshape(-1,1)))
+        results["slowness_factor"] = continuous_threshold(moving_avg(np.abs(vel_f), 20), thresh=0.06, steepness=6)
         for mod in model:
             if mod == "const_fric":
                 static_direction = np.sign(vel_f)
-                X = np.hstack((X, ((1 - slowness_factor_continuous(vel_f)) * static_direction).reshape(-1,1)))
+                X = np.hstack((X, ((1 - results["slowness_factor"]) * static_direction).reshape(-1,1)))
             elif mod == "const_opposing_fric":
                 opposing_direction = np.where(np.sign(vel_f) == np.sign(V_f), np.sign(vel_f), 0)
-                X = np.hstack((X, ((1 - slowness_factor_continuous(vel_f)) * opposing_direction).reshape(-1,1)))
+                X = np.hstack((X, ((1 - results["slowness_factor"]) * opposing_direction).reshape(-1,1)))
             # expect param["static_fric"] + param["V"] > 0 or else friction reverses
             elif mod == "static_fric":
-                X = np.hstack((X, (slowness_factor_continuous(vel_f) * V_f).reshape(-1,1)))
+                X = np.hstack((X, (results["slowness_factor"] * V_f).reshape(-1,1)))
             elif mod == "offset":
-                X = np.hstack((X, (1 - slowness_factor_continuous(vel_f)).reshape(-1,1)))
+                X = np.hstack((X, (1 - results["slowness_factor"]).reshape(-1,1)))
             elif mod == "Vsq":
                 X = np.hstack((X, (np.abs(V_f) * V_f).reshape(-1,1)))
             elif mod == "theta_3dot":   # alternate strategy to modeling current
@@ -213,59 +191,8 @@ def loadRun(filename):
     except Exception as e:
         print("Load failed:", filename, " because \n", e)
 
-def cut_first(data, filter_params):
-    return data[filter_params.N * 4:]
-
-def filter_data(data, filter_params=FilterParams(), type='lowpass'):
-    n_suspicious_filtered = filter_params.N * 2
-    padded_data = np.concatenate((np.average(data[:n_suspicious_filtered]) * np.ones(n_suspicious_filtered), data))
-    if type == 'lowpass' or type == 'low':
-        filtered = signal.sosfilt(filter_params.lowpass_sos, padded_data)
-    else:
-        filtered = signal.sosfilt(filter_params.hipass_sos, padded_data)
-    filtered = filtered[n_suspicious_filtered:]
-    if type == 'lowpass' or type == 'low':
-        filtered = cut_first(filtered, filter_params)
-    return filtered
-
 
 def clean_up_test_data(this_data, results, filter_params):
-    """ the angle is fuzzy, with noticeable spikes down from a probable true value. Remove those.
-    down: [True] to remove downspikes only, [True, False] to remove down then up for each width.
-    recalc change is whether to recalculate changes after every spike removal. might remove this."""
-    def remove_spikes(angle: np.ndarray, down: list[bool], recalc_change=False,
-                      min_spike_start_size=0.001, max_up_spike_width=4, max_down_spike_width=4):
-        future_length = max(max_up_spike_width, max_down_spike_width) + 1
-        prev_angle = np.roll(angle, 1)
-        prev_angle[0] = prev_angle[1]
-        # distance from previous angle to this angle and several future angles. [:,i] is i samples in the future.
-        change_from_prev_angle = np.empty((len(angle), future_length))
-        if not recalc_change:
-            for i in range(future_length):
-                change_from_prev_angle[:, i] = np.roll(angle, -i) - prev_angle
-        for width in range(1, future_length):
-            for is_down in down:
-                if width > (max_down_spike_width if is_down else max_up_spike_width):
-                    continue
-                if recalc_change:
-                    for i in range(future_length):
-                        change_from_prev_angle[:, i] = np.roll(angle, -i) - prev_angle
-                change_across_spike = change_from_prev_angle[:, width]
-                if is_down:
-                    is_away_from_prev = change_from_prev_angle < -min_spike_start_size
-                    ends_going_back = (change_from_prev_angle[:, width-1] < change_across_spike)
-                else:
-                    is_away_from_prev = change_from_prev_angle > min_spike_start_size
-                    ends_going_back = (change_from_prev_angle[:, width-1] > change_across_spike)
-                all_samples_away_from_prev = np.all(is_away_from_prev[:, 0:width], axis=1)
-                is_start_of_spike = (all_samples_away_from_prev & ends_going_back)
-                for index_in_spike in range(width):
-                    is_this_index_in_spike = np.roll(is_start_of_spike, index_in_spike)
-                    interpolated_value_unshifted = (index_in_spike + 1) / (width + 1) * change_across_spike + prev_angle
-                    interpolated_value = np.roll(interpolated_value_unshifted, index_in_spike)
-                    angle = np.where(is_this_index_in_spike, interpolated_value, angle)
-        return angle
-
     def remove_idle_messages(this_data, t_max=100):
         index_first = np.flatnonzero(this_data["t"] < t_max)[0]
         for m in ["V", "angle", "t"]:
@@ -282,6 +209,7 @@ def clean_up_test_data(this_data, results, filter_params):
         results["spiky_angle"] = np.concatenate((results["spiky_angle"], cut_first(this_data["angle"], filter_params)))
     else:
         results["spiky_angle"] = cut_first(this_data["angle"], filter_params)
+    # the angle is fuzzy, with noticeable spikes down from a probable true value. Remove those.
     this_data["angle"] = remove_spikes(this_data["angle"], [False], max_up_spike_width=2)
     this_data["angle"] = remove_spikes(this_data["angle"], [True], max_down_spike_width=4)
     # this_data["angle"] = remove_spikes(this_data["angle"], [False, True], recalc_change=True)
@@ -290,19 +218,6 @@ def clean_up_test_data(this_data, results, filter_params):
     this_data["angle"] = moving_avg(this_data["angle"], n_samples=FILT_MOVING_AVG_N, pad_val=this_data["angle"][0])
     duration = time.time() - startTime
     results["data clean time"] = results["data clean time"] + duration if "data clean time" in results else duration
-
-
-def moving_avg(data, n_samples=1, pad_val=0):
-    return np.concatenate((np.ones(n_samples-1) * pad_val, np.convolve(data, np.ones(n_samples), 'valid') / n_samples))
-
-# continuous number [0,1] where 1 means it's slow enough for static friction
-def slowness_factor_continuous(vel, speed_thresh=0.06, steepness=6, moving_avg_samples=20):
-    # arctangent is useful because it has a flat response in both limits
-    # return np.arctan((np.abs(vel) < speed_thresh) * 20 - 10) / np.pi + 0.5
-    # return np.arctan(speed_thresh / np.abs(vel)) / np.pi * 2
-    speed = moving_avg(np.abs(vel), moving_avg_samples)
-    return np.arctan((speed_thresh - speed) * steepness / speed_thresh) / np.pi + 0.5
-    # return np.clip(speed_thresh / np.abs(vel), 0., 1.)
 
 
 def saveParams(params, filename = "new"):
